+++ food-order/Entrypoint/Rest/Json/OrderResponse.cs
@@ -5,9 +5,9 @@ namespace food_order.Entrypoint.Rest.Json
         public string uuid { get; set; }
         public string restaurantUuid { get; set; }
 
-        public double total { get; set; }
+        public decimal total { get; set; }
 
-        public OrderResponse(string uuid, string restaurantUuid, double total) {
+        public OrderResponse(string uuid, string restaurantUuid, decimal total) {
           this.uuid = uuid;
           this.restaurantUuid = restaurantUuid;
           this.total = total;
